Extract `BaseOperator` class from (Workflow) `Operator`
