+++ nvtabular/graph/__init__.py
@@ -15,4 +15,5 @@
 #
 
 # flake8: noqa
+from .base_operator import BaseOperator
 from .node import Node
